args_parser: change URL handling
amir20@ reported a strange behaviour when an attempt to load test
https://www.clashleaders.com/ site is made. The responses returned
had 301 HTTP code, which seemed odd, since the URL provided is
correct and there shouldn't be a redirect. Turns out this was because
bombardier appends default port (80 for HTTP and 443 for HTTPS) to the
URLs used to send the requests, if no port is present.
A switch to github.com/goware/urlx was made, since standard library's
net/url package chose to interpret URLs in a slightly strange manner.
This solves the issue of appending unnecessary default ports when no
port is specified in the URL provided to bombardier, but also
introduces a breaking change in how the input is interpreted, so the
next version of the binary will have to be v2.0.0
The change is that some malformed URLs that were previously considered
to be valid are now not permitted - ":8080", "https://", "https://:443"
and such.